fix "RopeSlice to str conversion failed"
Initially, `RopeSlice.as_str()` was used because it doesn't create a
copy of the string, but it can fail if the `RopeSlice` is not continuous
in memory. So, we convert the `RopeSlice` to a string, because this will
not throw an error for large lines, and the `references` function is not
called often enough for us to care about copying one line.