feat: references, completion for type casting (#9)
* feat: completion for type casting

add completion for built-in function `<type>::into(input)`
also refactor `completion::types::FunctionTemplate`, removed redundant
field `snippet_base` and add new field `snippet`.

* add check for `into` function, fixed docs

using completion with function kind makes unpleasant overlap with word
`into` and it structure `<Input>::into(input)`, so we used snippet
completion kind for clean visual effect.

* change hover description

Include in description either it built-in or jet before function
signature. Also add after signature `---` for clean separation with it's
description.

* add `references` feature for functions

* remove unnecessary change for positions inside `find_all_references`

* fix "RopeSlice to str conversion failed"

Initially, `RopeSlice.as_str()` was used because it doesn't create a
copy of the string, but it can fail if the `RopeSlice` is not continuous
in memory. So, we convert the `RopeSlice` to a string, because this will
not throw an error for large lines, and the `references` function is not
called often enough for us to care about copying one line.

* add "reference" feauture for any calls

* clippy: remove unused import

* make definition show itself on reference

also change goto-definition to return range of symbol if we stay on
definition, because this actually triggers reference